// src/app/types/menu.ts
export type HeaderItem = {
  label: string;
  href: string;
};

export type avatar = {
  image: string;
  title: string;
};

export type innovation = {
  image: string;
  title: string;
  description: string;
  bg_color: string;
  txt_color: string;
};

export type WebResultTag = {
  image: string;
  name: string;
  bg_color: string;
  txt_color: string;
}

export type faq = {
  faq_que: string;
  faq_ans: string;
}